Add Promela and C++ examples
Contributes Promela models of a number of mutex and condition variable
implementations based on futexes, as well as Promela models for the
futex_wait and futex_wake variants of the system call. Also included
are C++ files showcasing implementations of futex-based mutexes from
Drepper's paper.

This code was co-authored with Hugues Evrard.